+++ logger.go
@@ -9,14 +9,19 @@ const (
 	debug = "DEBUG"
 )
 
-// Log conditionally STDOUTs the given message,
-// depending on DEBUG flag.
+// Log conditionally STDOUTs the given message, depending on DEBUG flag.
 func Log(msg ...string) {
 	if os.Getenv(debug) == "" {
 		return
 	}
 
-	for _, m := range msg {
-		fmt.Println(m)
+	for i, m := range msg {
+		fmt.Printf(m)
+
+		if i+1 != len(msg) {
+			fmt.Printf(" ")
+		}
 	}
+
+	fmt.Println("")
 }